Why: Restore CI/CD portability and replace the clearance-provider no-op with verified behavior.
Changed:
Implement pad/pad, pad/via, and via/via clearance diagnostics; add focused regression test; make model forward declarations MSVC-clean; make math constants portable; add explicit numeric conversions; preserve literal POSIX text-variable tokens in CLI integration tests; update handover docs.

Behavior:
Linux CLI tests retain ${VAR} tokens; MSVC warnings-as-errors no longer reject model headers or math code; clearance provider reports different-net proximity violations.

Verification:
MinGW focused CLI and clearance tests passed; full Qt build 112/112; CTest 75/75; official visual harness sprint391_ci_portability_proof passed; screenshot inspected; GUI stderr empty.

Demo:
artifacts/screenshots/sprint391_ci_portability_proof-20260916-102638.png

+++ src/ccad_core/drc.cpp
@@ -1600,8 +1600,8 @@ void checkSolderMaskBridges(const Board& board, std::vector<Diagnostic>& diagnos
   auto maskCorners = [&board](const Pad& pad) {
     std::vector<Point> corners = padCorners(pad);
     const long double expansion = static_cast<long double>(board.design_rules.solder_mask_expansion.nanometers);
-    long double min_x = corners.front().x.nanometers, max_x = min_x;
-    long double min_y = corners.front().y.nanometers, max_y = min_y;
+    long double min_x = static_cast<long double>(corners.front().x.nanometers), max_x = min_x;
+    long double min_y = static_cast<long double>(corners.front().y.nanometers), max_y = min_y;
     for (const Point& point : corners) {
       min_x = std::min(min_x, static_cast<long double>(point.x.nanometers));
       max_x = std::max(max_x, static_cast<long double>(point.x.nanometers));
